monitoring: Share parse_start_end helper
